Move common setup into base generator module
