+++ Gemfile
@@ -8,3 +8,5 @@ gemspec
 gem "rake", "~> 13.0"
 
 gem "colorize", "~> 0.8"
+
+gem "byebug"
