+++ lib/generators/rails_template18f/github_actions/github_actions_generator.rb
@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 
+require "rails/generators"
+
 module RailsTemplate18f
   module Generators
     class GithubActionsGenerator < ::Rails::Generators::Base
-      include ::Rails::Generators::AppName
-      include RailsTemplate18f::TerraformOptions
+      include Base
+      include TerraformOptions
 
       class_option :node_version, desc: "Node version to test against in actions"
 
@@ -12,10 +14,6 @@ module RailsTemplate18f
         Description:
           Install Github Actions workflow files
       DESC
-
-      def self.source_root
-        @source_root ||= File.expand_path("templates", __dir__)
-      end
 
       def install_actions
         directory "github", ".github"
