+++ lib/rails_template_18f.rb
@@ -1,12 +1,9 @@
 # frozen_string_literal: true
 
 require_relative "rails_template18f/version"
+require_relative "rails_template18f/generators"
 
 module RailsTemplate18f
-  extend ActiveSupport::Autoload
-
-  autoload :TerraformOptions
-
   class Error < StandardError; end
 
   class Railtie < ::Rails::Railtie; end
